Rename: 폴더 구조를 정리한다.
