+++ client/src/commons/atoms/buttons/revise-remove/RemoveBtn.tsx
@@ -1,6 +1,6 @@
 import { Styledprops } from '@/types';
-import { RecuitBtn } from '../../header/CHeader.styled';
+import { RecuitBtn } from '../../../../components/header/CHeader.styled';
 
-export default function RemoveBtn( props:Styledprops ) {
+export default function RemoveBtn(props: Styledprops) {
   return <RecuitBtn {...props}>삭제</RecuitBtn>;
 }
