+++ client/src/commons/atoms/buttons/revise-remove/ReviseBtn.tsx
@@ -1,5 +1,5 @@
 import { Styledprops } from '@/types';
-import { RecuitBtn } from '../../header/CHeader.styled';
+import { RecuitBtn } from '../../../../components/header/CHeader.styled';
 
 export default function ReviseBtn(props: Styledprops) {
   return <RecuitBtn {...props} >수정</RecuitBtn>;
