+++ client/src/commons/molecules/profile/MemberProfile.styled.tsx
@@ -1,0 +1,8 @@
+import { styled } from 'styled-components';
+import tw from 'twin.macro';
+
+import { FlexWrapper } from '@/commons/styles/Containers.styled';
+
+export const MemberProfileWrapper = styled(FlexWrapper)`
+  ${tw`items-center`}
+`
