+++ client/src/commons/molecules/profile/MemberProfile.tsx
@@ -1,10 +1,11 @@
 /* 2023-07-03 유저 프로필 컴포넌트 - 김다함 */
 import { Member } from '@/types';
 
-import { FlexColumnWrapper, FlexWrapper } from '../styles/Containers.styled';
-import { SmallText } from '../atoms/Typography';
-import Image from '@/commons/atoms/Image';
-import Label from '@/commons/atoms/Label';
+import { MemberProfileWrapper } from '@/commons/molecules/profile/MemberProfile.styled';
+import { FlexColumnWrapper } from '@/commons/styles/Containers.styled';
+import { SmallText } from '@/commons/atoms/text/Typography';
+import Image from '@/commons/atoms/image/Image';
+import Label from '@/commons/atoms/text/Label';
 
 interface MemberProfile {
   type: 'board' | 'comment' | 'portfolio' | 'blackboard';
@@ -21,17 +22,18 @@ const ImageSizes: any = {
 
 const MemberProfile = ({ type, member, date }: MemberProfile) => {
   return (
-    <FlexWrapper gap={15} className='items-center'>
+    <MemberProfileWrapper gap={15}>
       <Image src={member.picture} url={`/members/${member.memberId}`} shape='circle' size={ImageSizes[type]} />
-      {type === 'portfolio' ?
+      {type === 'portfolio' &&
         <Label text={member.name} type={type} url={`/members/${member.memberId}`} />
-        :
-        <FlexColumnWrapper gap={0}>
+      }
+      {type !== 'portfolio' &&
+        <FlexColumnWrapper>
           <Label text={member.name} type={type} url={`/members/${member.memberId}`} />
-          <SmallText className={type === 'blackboard' ? 'text-white' : ''}>{date}</SmallText>
+          <SmallText color={type === 'blackboard' ? 'white' : ''}>{date}</SmallText>
         </FlexColumnWrapper>
       }
-    </FlexWrapper >
+    </MemberProfileWrapper >
   )
 }
