+++ client/src/commons/styles/Containers.styled.tsx
@@ -1,9 +1,9 @@
-/* 2023.07.03 컴포넌트, 페이지에 전체적으로 쓰이는 정렬 컨테이너 - 김다함 */
+/* 2023.07.03 모든 컴포넌트, 페이지에 전체적으로 쓰이는 정렬 컨테이너 - 김다함 */
 import { styled } from 'styled-components';
 import tw from 'twin.macro';
 
 const Container = styled.div<{ gap?: number, bg?: string }>`
-  ${tw`flex`}
+  ${tw`flex flex-wrap`}
   gap: ${(props) => props.gap || 0}px;
   background-color: ${(props) => props.bg || 'white'};
 `
@@ -29,7 +29,10 @@ export const FlexWrapper = styled.div<{ gap?: number }>`
   gap: ${(props) => props.gap}px;
 `;
 
-export const FlexColumnWrapper = styled.div<{ gap?: number }>`
-  ${tw`flex flex-col `}
-  gap: ${(props) => props.gap}px;
+export const FlexBetweenWrapper = styled(FlexWrapper)`
+  ${tw`justify-between`}
+`;
+
+export const FlexColumnWrapper = styled(FlexWrapper)`
+  ${tw`flex-col`}
 `;
