--- client/src/commons/styles/StyleList.ts
@@ -1,14 +1,0 @@
-/* 2023-07-04 디자인 토큰 대신 사용할까 고민중이나 아직 불필요해서 보류 - 김다함*/
-import { Profile } from '@/types/style';
-
-export const profileTypes: Profile = {
-  comment: {
-    fontSize: '15px',
-  },
-  board: {
-    fontSize: '20px',
-  },
-  portfolio: {
-    fontSize: '40px',
-  },
-}
