+++ client/src/commons/styles/layout/MainLayout.tsx
@@ -1,5 +1,5 @@
-import Header from '../../atoms/header/Header';
-import Footer from '../../atoms/footer/Footer';
+import Header from '../../../components/header/Header';
+import Footer from '../../../components/footer/Footer';
 import { Outlet } from 'react-router-dom';
 import { MainImgWrapper } from './Layout.styled';
