+++ client/src/components/editor/Editor.styled.tsx
@@ -2,6 +2,33 @@
 import { styled } from 'styled-components';
 import tw from 'twin.macro';
 
+export const QuillWrapper = styled.div<{ isTitleFormOpen: boolean }>`
+  z-index: 10;
+  position: absolute;
+  border: 0;
+  top: 0;
+  pointer-events: ${({ isTitleFormOpen }) => isTitleFormOpen && 'none'};
+  .ql-toolbar {
+    padding: 17px 0;
+    background-color: #252525;
+    border: 0;
+    text-align: center;
+  }
+  .ql-stroke, .ql-fill {
+    stroke: white;
+  }
+  .ql-picker-label {
+    color: white;
+  }
+  .ql-container{
+    border: 0;
+  }
+  .ql-editor{
+    width: 1000px;
+    height: calc(100vh - 60px);
+  }
+`
+
 export const Input = styled.input`
     ${tw`w-full border-0`}
 `;
@@ -28,3 +55,9 @@ export const DarkTextArea = styled(TextArea)`
     background-color: #3A3B41;
     color: white;
 `;
+
+export const TitleFormContainer = styled.div`
+      ${tw`w-screen px-16 py-7 rounded-t-2xl z-20 bottom-0 absolute`};
+      background-color: #161616;
+      box-shadow: 0 -8px 10px -1px #a9a9a9;
+  `;
