+++ client/src/components/editor/QuillEditor.tsx
@@ -1,38 +1,12 @@
 /* 2023-07-06 React Quill Editor - 김다함 */
 import { memo, useMemo, useRef } from 'react';
-import { styled } from 'styled-components';
 import "react-quill/dist/quill.snow.css";
 import ReactQuill from 'react-quill';
 
-import { Quill } from '@/types';
 import useImageHandler from '@/hooks/useImageHandler';
 
-const QuillWrapper = styled.div<{ isTitleFormOpen: boolean }>`
-  z-index: 10;
-  position: absolute;
-  border: 0;
-  top: 0;
-  pointer-events: ${({ isTitleFormOpen }) => isTitleFormOpen && 'none'};
-  .ql-toolbar {
-    padding: 17px 0;
-    background-color: #252525;
-    border: 0;
-    text-align: center;
-  }
-  .ql-stroke, .ql-fill {
-    stroke: white;
-  }
-  .ql-picker-label {
-    color: white;
-  }
-  .ql-container{
-    border: 0;
-  }
-  .ql-editor{
-    width: 1000px;
-    height: calc(100vh - 60px);
-  }
-`
+import { QuillWrapper } from '@/components/editor/Editor.styled';
+import { Quill } from '@/types';
 
 const QuillEditor = memo(({ htmlContent, setContentHandler, isTitleFormOpen }: Quill) => {
   const [imageUrlHandler, imageHandler] = useImageHandler();
