+++ client/src/components/footer/Footer.tsx
