+++ client/src/pages/community-detail/CommunityDetail.tsx
@@ -2,10 +2,10 @@ import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { call } from '@/utils/apiService';
 
-import MemberProfile from '@/commons/molecules/MemberProfile';
+import MemberProfile from '@/commons/molecules/profile/MemberProfile';
 import DetailContents from '@/components/detailContents/DetailContents';
 import Loading from '../../components/loading/Loading'; // components/loding 이동 (page x) 수정 완 
-import CommentBox from '@/components/CommentBox';
+import CommentBox from '@/components/comment/CommentBox';
 import DeleteModal from '@/components/modal/DeleteModal';
 
 import { CommuProps } from '@/types';
