+++ client/src/pages/portfolio-detail/PortfolioDetail.tsx
@@ -8,17 +8,17 @@ import { Portfolio, Member, Tag as tag } from '@/types';
 import { call } from '@/utils/apiService';
 
 import { ButtonHeader, ContentContainer, PortfolioContainer, HtmlWrapper, UserCard, UserContainer } from './PortfolioDetail.styled';
-import { Center, FlexColumnContainer, FlexWrapper } from '@/commons/styles/Containers.styled';
-import { BodyText, HeadingText, LabelText, SmallText } from '@/commons/atoms/Typography';
+import { Center, FlexBetweenWrapper, FlexColumnContainer, FlexWrapper } from '@/commons/styles/Containers.styled';
+import { BodyText, HeadingText, LabelText, SmallText } from '@/commons/atoms/text/Typography';
 import ReviseBtn from '@/commons/atoms/buttons/revise-remove/ReviseBtn';
 import RemoveBtn from '@/commons/atoms/buttons/revise-remove/RemoveBtn';
 import { AskCommisionBtn } from '@/commons/atoms/buttons/Button.styled';
-import MemberProfile from '@/commons/molecules/MemberProfile';
+import MemberProfile from '@/commons/molecules/profile/MemberProfile';
 import DeleteModal from '@/components/modal/DeleteModal';
 import Bookmark from '@/commons/atoms/buttons/Bookmark';
 import LikeBtn from '@/commons/atoms/buttons/LikeBtn';
 import { BsArrowReturnLeft } from 'react-icons/bs';
-import Tag from '@/commons/molecules/Tag';
+import Tag from '@/commons/molecules/tag/Tag';
 
 
 export default function PortfolioDetail() {
@@ -67,7 +67,7 @@ export default function PortfolioDetail() {
 
         <UserContainer>
           <UserCard>
-            <FlexWrapper className='justify-between'>
+            <FlexBetweenWrapper>
               {portfolio &&
                 <>
                   <LikeBtn portfolioId={portfolio.portfolioId} currentLikes={portfolio.likes} isToggled={portfolio.isLiked} />
@@ -77,7 +77,7 @@ export default function PortfolioDetail() {
                   </FlexWrapper>
                 </>
               }
-            </FlexWrapper>
+            </FlexBetweenWrapper>
             {member &&
               <MemberProfile type="portfolio" member={member} />
             }
