+++ client/src/pages/portfolio-edit/PortfolioEdit.styled.tsx
@@ -1,4 +1,3 @@
-/* 2023-07-05 홈페이지 내 들어가는 모든 버튼 styled - 김다함*/
 import { styled, css } from 'styled-components';
 import tw from 'twin.macro';
 
@@ -31,4 +30,8 @@ export const PortfolioEditButton = styled.button<{ color: 'light' | 'dark' | 'bl
         border: 1px solid black;
       `
   }
+`
+
+export const PortfolioCheckButton = styled(PortfolioEditButton)`
+  ${tw`absolute bottom-10 right-16`};
 `
